+++ source/extensions/octi.lab_tasks/octi/lab_tasks/cfgs/robots/leap_hand/__init__.py
@@ -1,3 +1,4 @@
 from .robot_cfg import (
+    LEAP_DEFAULT_JOINT_POS,
     IMPLICIT_LEAP
 )