+++ source/extensions/octi.lab_tasks/octi/lab_tasks/tasks/manipulations/__init__.py
@@ -1,1 +1,2 @@
 from .track_goal import *
+from .lift_cube import *