+++ source/extensions/octi.lab_tasks/octi/lab_tasks/tasks/manipulations/lift_cube/config/__init__.py
@@ -1,1 +1,2 @@
 from .hebi import *
+from .leap_hand_xarm import *