+++ source/extensions/octi.lab_tasks/octi/lab_tasks/tasks/manipulations/track_goal/config/__init__.py
@@ -1,2 +1,3 @@
 from .hebi import *
 from .leap_hand import *
+from .leap_hand_xarm import *